+++ claude_py_scaffold/handlers.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, Request, status
-from fastapi.responses import JSONResponse
 from fastapi.exceptions import RequestValidationError
+from fastapi.responses import JSONResponse
 from sqlalchemy.exc import SQLAlchemyError
 
 from claude_py_scaffold.exceptions import AppException